added user permissions

+++ main/views.py
@@ -25,39 +25,63 @@ def detail(request, id):
 
 # add movies to the database
 def add_movies(request):
-    if request.method == 'POST':
-        form = MovieForm(request.POST or None)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            if request.method == 'POST':
+                form = MovieForm(request.POST or None)
 
-        if form.is_valid():
-            data = form.save(commit=False)
-            data.save()
+                if form.is_valid():
+                    data = form.save(commit=False)
+                    data.save()
+                    return redirect('main:home')
+            else:
+                form = MovieForm()
+            return render(request, 'main/addmovies.html', {'form': form,
+                                                           'controller': 'Add Movies'})
+        # if they are not admin
+        else:
             return redirect('main:home')
-    else:
-        form = MovieForm()
-    return render(request, 'main/addmovies.html', {'form': form,
-                                                   'controller': 'Add Movies'})
+
+    # if they are not logged in
+    return redirect('accounts:login')
 
 
 # edit the movie
 def edit_movies(request, id):
-    movie = Movie.objects.get(id=id)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            movie = Movie.objects.get(id=id)
 
-    if request.method == 'POST':
-        form = MovieForm(request.POST or None, instance=movie)
+            if request.method == 'POST':
+                form = MovieForm(request.POST or None, instance=movie)
 
-        if form.is_valid():
-            data = form.save(commit=False)
-            data.save()
-            return redirect('main:detail', id)
-    else:
-        form = MovieForm(instance=movie)
-    return render(request, 'main/addmovies.html', {'form': form,
-                                                   'controller': 'Edit Movies'})
+                if form.is_valid():
+                    data = form.save(commit=False)
+                    data.save()
+                    return redirect('main:detail', id)
+            else:
+                form = MovieForm(instance=movie)
+            return render(request, 'main/addmovies.html', {'form': form,
+                                                           'controller': 'Edit Movies'})
+        # if they are not admin
+        else:
+            return redirect('main:home')
+
+    # if they are not logged in
+    return redirect('accounts:login')
 
 
 # delete movies
 def delete_movies(request, id):
-    movie = Movie.objects.get(id=id)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            movie = Movie.objects.get(id=id)
 
-    movie.delete()
-    return redirect('main:home')
+            movie.delete()
+            return redirect('main:home')
+        # if they are not admin
+        else:
+            return redirect('main:home')
+
+    # if they are not logged in
+    return redirect('accounts:login')
